--- a/.gitignore
+++ b/.gitignore
@@ -7,4 +7,12 @@
 .vscode/
 *.log
 **.ipynb_checkpoints/
+.venv
+build/
+dist/
+ocelot.egg-info/
 .DS_Store
+arcline_traj.png
+arcline_traj_0.png
+dev_tests/5_CSR.py
+dev_tests/__init__.py