… ocelot/cpbd/optics.py