Hi everyone,

I have been working for some time on a few changes which can significantly speed-up particle tracking (with transfer maps) and the calculation of CSR effects. Many of these changes come directly from my own small tracking code [`Wake-T`](https://github.com/AngelFP/Wake-T) for plasma accelerators, in which the transfer maps and CSR are anyways based on the `Ocelot` implementation.

## Changes
The changes are only in the `optics.py` and `csr.py` files and can be summarized as follows:
* `optics.py` 
    * The `numpy_apply` method has been greatly simplified to a single line using 'np.einsum'. It now performs full matrix multiplication and is also ~20-30% faster than before.
    * The `numexpr_apply` method has also been modified in order not to require making a copy of the initial beam matrix and to simplify the creation of the beam arrays and Rij elements. As a result, this method is also ~20-30% faster than before.
    * The `numba_apply` method has only a few cosmetic changes and the `parallel=True` argument has been removed, since it didn't work anyways (at least in my tests).
    * In my tests, the `numba_apply` method was always slightly faster than `numpy_apply`, so I have changed the priority such that `numba_apply` is used if `numba` is installed.
    * Since now both `numpy_apply` and `numba_apply` have full matrix multiplication, I have also changed how the `self.full_matrix` parameter is used.

* `csr.py`:
    The changes with the biggest impact are those to the methods for computing `K0`. These are the following:
    * The implementation of `K0_fin_anf_np` is now significantly more efficient by changing the way in which some arrays are generated and computed. Redundant indices `i1` and `idx1` have been removed. Some of these changes have also been applied to `K0_fin_anf_numexpr` and `K0_fin_anf_opt`.
    * A new method `estimate_start_index` has been added which is used by `K0_fin_anf_np` and `K0_fin_anf_numexpr`. This method can significantly speed up the calculation of CSR effects by allowing to pre-discard some regions of the reference trajectory which do not contribute to the CSR kernel (see method documentation for more information).
    * The numba method `K0_0_jit` has also been modified to stop the loop once it detects that the trajectory points no longer contribute to the CSR kernel (i.e, once `w <= wmin`). This is more precise than the `estimate_start_index` method used by `numpy` and `numexpr`. Previously, the `n` and `R` arrays were computed for the whole trajectory, and only afterwards the parts where `w <= wmin` would be removed. This is now no longer necessary and only the required points are computed, leading to a potential speed-up of the calculation. The methods `K0_fin_anf_opt` and `K0_1_jit` have been adapted to take these changes into account.

    Other changes are:
    * I have removed the unused methods `convolution` and `apply_i`, as well as `interp1`, which has been replaced by `np.interp` because that's what it was always calling anyways.
    * I have changed a bit the imports to better organize them and avoid using `*`.
    * Simplified loop in `apply`.
    * Removed `w, idx = np.unique(w, return_index=True)` in `CSR_K1` because, if I understood it well, it might not be needed if `w[0]` is replaced by `np.min(w)` in the following lines. However, please correct this if I'm wrong.
    * Some formatting changes here and there.

## Benchmarks
To test the increase in performance and check that the physics is still correct, I have run the CSR example from the IPython tutorials. These were the results (units are in seconds):
| Using |  Before (no CSR)  |  After  (no CSR) |  Before (with CSR)  |  After  (with CSR)   |
|----|---|---|---|---|
|  `numpy`  |  2.6 | 2.0  | 170.3  |  99.6 |
|  `numpy` + `numba`   | 2.7  | 1.8  | 95.6  | 76.2  |
|  `numpy` + `numba` + `numexpr` | 0.7  | 0.5  |  92.7 | 70.9  |

While the values without CSR are usually quite consistent, the ones with CSR tend to fluctuate by up to ~10s from run to run. In any case, the results are always identical (at least to my eye), producing always:
![arcline_traj](https://user-images.githubusercontent.com/20479420/82228070-ef415800-9928-11ea-8828-23545083d67c.png)

Therefore, from these results and from how I understand the code, none of the changes have an impact on the physics. However, if you have other cases which can be simulated to validate the results, I would still recommend to test them.

I hope you find these changes useful, but feel free to suggest any modifications.