@sergey-tomin 2020-05-27
Hi Ángel,
Thank you for your contribution. I'm looking at your pull request. I ran init tests and a few unit tests failed. The reason seems to be a modification of the CSR module. Although the difference is small I need to understand where it comes from and can it be a potential problem in the future. So, it will take a some time.