@AngelFP 2020-05-28
Hi Sergey,

Thank you for having a look at it. If there are any tests I can run myself, you can point me to them and I can try to see where these differences might be coming from.