@sergey-tomin 2020-05-28
Hi Ángel,
just have a look to https://github.com/ocelot-collab/ocelot/tree/master/unit_tests 
there is readme.txt with instructions how to run unit-tests.