@sergey-tomin 2020-06-03
Hi Ángel,

I went through the changes with optics.py and I did not have any problem. Actually, it is nice changes. 
For CSR it is a bit different story. I checked what you did and got idea of your changes. As I said earlier, the unit-tests regarding CSR are failed. I seem to have discovered where the difference comes from. in CSR class you modified methods "K0_fin_anf_...". In original, the last lines are ending with this:
 if len(K) > 1:
        a = np.append(0.5 * (K[0:-1] + K[1:]) * np.diff(s), 0.5 * K[-1] * s[-1])
        KS = np.cumsum(a[::-1])[::-1]
 else:
        KS = 0.5 * K[-1] * s[-1]
I would leave it as it is