@AngelFP 2020-06-04
Hi Sergey,

I'm happy to see it got solved! I didn't have any time yet to try the unit-tests, so thank you for having a detailed look at the code and finding where the issue was. I'm glad you find the changes useful!

By the way, when working on the code there was something I didn't fully understand regarding the calculation of the CSR kernel. It is done with two different types of functions: `k0_fin_anf` and `k0_fin_inf`. When looking at the literature about this 1D method, I found some papers from M. Dohlus where the function for calculating the kernel looks like the one in `k0_fin_anf`, but I couldn't find anything about `k0_fin_inf`.  My guess is that one is for the long-range interactions and the other for the short range, but I'm not really sure. Is this the case?