@sergey-tomin 2020-06-11
Hi Ángel,
I double check with Martin. It is part which is coming from the infinite straight line that is assumed before the specified CSR region and it is calculated analytically.