@AngelFP 2020-06-19
Hi Sergey,
I see I forgot to reply here! Thanks a lot for the information. I must say I learned quite a lot about the CSR model from working on the code!